Rename files to ts to preserve history
