+++ packages/redux-query-react/.eslintrc.js
@@ -28,6 +28,11 @@ module.exports = {
     react: {
       version: 'detect',
     },
+    'import/resolver': {
+      node: {
+        extensions: ['.js', '.jsx', '.ts', '.tsx'],
+      },
+    },
   },
   rules: {
     'react/prop-types': 0,
