+++ packages/redux-query-react/src/context.ts
@@ -1,7 +1,7 @@
 import React from 'react';
 
 type ReduxQueryContextValue = {
-  queriesSelector: (state: any, ...rest: any) => any,
+  queriesSelector: (state: any, ...rest: any) => any;
 };
 
 const ReduxQueryContext = React.createContext < ReduxQueryContextValue > null;
