+++ packages/redux-query/.eslintrc.js
@@ -18,4 +18,11 @@ module.exports = {
     jest: true,
     node: true,
   },
+  settings: {
+    'import/resolver': {
+      node: {
+        extensions: ['.js', '.jsx', '.ts', '.tsx'],
+      },
+    },
+  },
 };
